requirements.txt
fastapi
aiologger[aiofiles]
pip-chill
gunicorn
uvicorn
pydantic
sentence_transformers
feedparser
scikit-learn
numpy
aiocache
python-dotenv
requests
beautifulsoup4
lxml
anthropic
accelerate 
tiktoken
yandex-cloud-ml-sdk
python-dotenv
aiohttp